chore: clean up zoe swingset tests, split CI into "unit" (<1m) and "swingset" tests (> 6m)
