--- packages/zoe/test/swingsetTests/zoe-metering/infiniteInstallLoop.js
@@ -1,3 +1,0 @@
-for (;;) {
-  // Nothing.
-}
