--- packages/zoe/test/swingsetTests/zoe-metering/infiniteInstanceLoop.js
@@ -1,5 +1,0 @@
-export const start = () => {
-  for (;;) {
-    // Nothing.
-  }
-};
