--- packages/zoe/test/swingsetTests/zoe-metering/infiniteTestLoop.js
@@ -1,12 +1,0 @@
-import { Far } from '@agoric/marshal';
-
-export const start = _zcf => {
-  const publicFacet = Far('publicFacet', {
-    doTest: () => {
-      for (;;) {
-        // Nothing
-      }
-    },
-  });
-  return harden({ publicFacet });
-};
