--- packages/zoe/test/swingsetTests/zoe-metering/testBuiltins.js
@@ -1,10 +1,0 @@
-import { Far } from '@agoric/marshal';
-
-export const start = _zcf => {
-  const publicFacet = Far('publicFacet', {
-    doTest: () => {
-      new Array(1e9).map(Object.create);
-    },
-  });
-  return harden({ publicFacet });
-};
